fix(st-uart): skip console flush if UART is disabled
Check the USART_CR1_UE bit and if it is 0, the UART is not enabled,
or not clocked (but the read won't freeze the bus and will return 0).
In this case skip the console flush.
If flush didn't happen in time, just exit the function instead of panic.

Signed-off-by: Yann Gautier <yann.gautier@foss.st.com>
Change-Id: I5d1ef7e51612b4795e314b2f2da04a514b6c96a0
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/315467
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>